feat(review): add linguist time spent report

+++ webapp/src/main/java/com/box/l10n/mojito/service/review/ReviewProjectTimeSpentDecisionRow.java
@@ -3,4 +3,4 @@ package com.box.l10n.mojito.service.review;
 import java.time.ZonedDateTime;
 
 public record ReviewProjectTimeSpentDecisionRow(
-    Long decisionId, Long decisionUserId, ZonedDateTime decidedAt, Long wordCount) {}
+    Long decisionId, Long decisionUserId, ZonedDateTime decidedAt, Integer wordCount) {}
